Make ContextualData polyglot
Closes #1
